+++ Assets/lilToon/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.4] - 2022-07-xx
+### Fixed
+- Fixed an issue that could cause script errors
+- Fixed color value after optimization
+
 ## [1.3.3] - 2022-07-27
 ### Added
 - `UV Mode` for glitter mask